+++ CHANGELOG.md
@@ -1,5 +1,4 @@
 # Changelog
 
-* GUI
-* Sound effects
-* Ability to edit settings
+* An ability to turn the background music on and off was added
+* An ability to turn the sfx on and off was added